// src/Frontend/ViteApp/src/api-generated/models/LayoutItemType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KISv4 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const LayoutItemType = {
    SaleItem: 'SaleItem',
    Pipe: 'Pipe',
    Layout: 'Layout'
} as const;
export type LayoutItemType = typeof LayoutItemType[keyof typeof LayoutItemType];


export function instanceOfLayoutItemType(value: any): boolean {
    for (const key in LayoutItemType) {
        if (Object.prototype.hasOwnProperty.call(LayoutItemType, key)) {
            if (LayoutItemType[key as keyof typeof LayoutItemType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function LayoutItemTypeFromJSON(json: any): LayoutItemType {
    return LayoutItemTypeFromJSONTyped(json, false);
}

export function LayoutItemTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): LayoutItemType {
    return json as LayoutItemType;
}

export function LayoutItemTypeToJSON(value?: LayoutItemType | null): any {
    return value as any;
}

export function LayoutItemTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): LayoutItemType {
    return value as LayoutItemType;
}