// src/Frontend/ViteApp/src/api-generated/models/LayoutReadResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KISv4 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { LayoutItemModel } from './LayoutItemModel';
import {
    LayoutItemModelFromJSON,
    LayoutItemModelFromJSONTyped,
    LayoutItemModelToJSON,
    LayoutItemModelToJSONTyped,
} from './LayoutItemModel';

/**
 * 
 * @export
 * @interface LayoutReadResponse
 */
export interface LayoutReadResponse {
    /**
     * 
     * @type {number}
     * @memberof LayoutReadResponse
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof LayoutReadResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof LayoutReadResponse
     */
    image: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof LayoutReadResponse
     */
    topLevel: boolean;
    /**
     * 
     * @type {Array<LayoutItemModel>}
     * @memberof LayoutReadResponse
     */
    layoutItems: Array<LayoutItemModel>;
}

/**
 * Check if a given object implements the LayoutReadResponse interface.
 */
export function instanceOfLayoutReadResponse(value: object): value is LayoutReadResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('image' in value) || value['image'] === undefined) return false;
    if (!('topLevel' in value) || value['topLevel'] === undefined) return false;
    if (!('layoutItems' in value) || value['layoutItems'] === undefined) return false;
    return true;
}

export function LayoutReadResponseFromJSON(json: any): LayoutReadResponse {
    return LayoutReadResponseFromJSONTyped(json, false);
}

export function LayoutReadResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): LayoutReadResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'image': json['image'],
        'topLevel': json['topLevel'],
        'layoutItems': ((json['layoutItems'] as Array<any>).map(LayoutItemModelFromJSON)),
    };
}

export function LayoutReadResponseToJSON(json: any): LayoutReadResponse {
    return LayoutReadResponseToJSONTyped(json, false);
}

export function LayoutReadResponseToJSONTyped(value?: LayoutReadResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'image': value['image'],
        'topLevel': value['topLevel'],
        'layoutItems': ((value['layoutItems'] as Array<any>).map(LayoutItemModelToJSON)),
    };
}