// src/Frontend/ViteApp/src/api-generated/models/PipeListModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KISv4 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CashBoxesReadIdParameter } from './CashBoxesReadIdParameter';
import {
    CashBoxesReadIdParameterFromJSON,
    CashBoxesReadIdParameterFromJSONTyped,
    CashBoxesReadIdParameterToJSON,
    CashBoxesReadIdParameterToJSONTyped,
} from './CashBoxesReadIdParameter';

/**
 * 
 * @export
 * @interface PipeListModel
 */
export interface PipeListModel {
    /**
     * 
     * @type {CashBoxesReadIdParameter}
     * @memberof PipeListModel
     */
    id: CashBoxesReadIdParameter;
    /**
     * 
     * @type {string}
     * @memberof PipeListModel
     */
    name: string;
}

/**
 * Check if a given object implements the PipeListModel interface.
 */
export function instanceOfPipeListModel(value: object): value is PipeListModel {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function PipeListModelFromJSON(json: any): PipeListModel {
    return PipeListModelFromJSONTyped(json, false);
}

export function PipeListModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): PipeListModel {
    if (json == null) {
        return json;
    }
    return {
        
        'id': CashBoxesReadIdParameterFromJSON(json['id']),
        'name': json['name'],
    };
}

export function PipeListModelToJSON(json: any): PipeListModel {
    return PipeListModelToJSONTyped(json, false);
}

export function PipeListModelToJSONTyped(value?: PipeListModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': CashBoxesReadIdParameterToJSON(value['id']),
        'name': value['name'],
    };
}